// lite-api-web/api/api/models/ApiInfoDo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiInfoDo = {
    id?: number;
    name?: string;
    description?: string;
    url?: string;
    requestParams?: string;
    responseParams?: string;
    responseExample?: string;
    requestHeader?: string;
    responseHeader?: string;
    status?: number;
    method?: string;
    userId?: number;
    createTime?: string;
    updateTime?: string;
    deleted?: number;
};